+++ src/boot/ksdk1.1.0/devSSD1331.h
@@ -44,7 +44,7 @@ typedef enum
 
 int	devSSD1331init(void);
 
-int pullingforceprint(int16_t y[], uint8_t length);
+uint8_t pullingforceprint(int16_t y[], uint8_t length);
 int writetoscreen(int ar[]);
 int writetoscreeupper(int upper);
 int writetoscreenlower(int lower);
