+++ src/boot/ksdk1.1.0/ergsystem.h
@@ -2,5 +2,5 @@
 
 int intsplittoarrayupper(int data);
 int intsplittoarraylower(int data);
-int reduce_accel_array_resize_offset(int16_t acc[], uint16_t length, uint16_t index);
+uint8_t reduce_accel_array_resize_offset(int16_t acc[], uint16_t length, uint16_t index);
 //int offset_av_calc(int length, int menuI2cPullupValue);
